refactor team routes
